Codify dependency on `EtOrbi::EoTime#to_local_time`
The code as written depends on the `EtOrbi::EoTime#to_local_time`
method. However, this method was not introduced until `et-orbi`
version `1.0.6`.

In cases where the bundle already contains a version of `et-orbi` that
is not updated to `1.0.6`, the code fails to execute because the
method does not exist yet.

fixes #4